feat: 편지 일러스트 고르는 페이지로 연결
