+++ rememb/src/pages/MyHome.js
@@ -1,8 +1,8 @@
 import React from "react";
 import MainTitle from "../components/CommonHome/MainTitle";
 import PartyRoom from "../components/CommonHome/PartyRoom";
-import ToBalance from './../ToBalance';
-import WatchBalance from './../WatchBalance';
+import ToBalance from '../components/MyHome/ToBalance';
+import WatchBalance from '../components/CommonHome/WatchBalance';
 
 const MyHome = () => {
   // const [who,setWho]=useState('');
