+++ rememb/src/pages/YourHome/Letter/SelectImg.js
@@ -1,0 +1,12 @@
+import MainTitle from "../../../components/CommonHome/MainTitle";
+
+const SelectImg = () => {
+    return (
+      <div>
+        <MainTitle/>
+        이미지를 고르기
+      </div>
+    );
+  };
+  
+  export default SelectImg;
